Update README.md
-	Added error handling to the following-
a)	The _safeMint function has been added to handle errors related to minting. It checks that the recipient address is not the zero address and that the token has not already been minted.
b)	The _exists function has been modified to avoid a recursive call and to handle errors related to token existence. It simply checks if the token exists based on its ID.
c)	The setTokenURI function has been modified to handle errors related to setting the token URI. It now checks if the token exists before allowing the URI to be set.
d)	The updatePrice function has been modified to handle errors related to updating the token price. It checks if the token exists before allowing the price to be updated.

These error handling measures help prevent potential issues such as transferring funds without a positive balance or to an invalid recipient address.
-	Corrected an syntax error by replacing pragma solidity ^0.8.0 with pragma solidity  ^0.8.0;.
-	Added relevant comments to the code to each function and section of the code to explain their purpose and functionality.
-	Applied code indentation for neatness.
-	Added punctuation marks wherever, required.
-	Did paraphrasing.
-	Corrected grammatical errors.
-	Formatted the entire document by adding/removing adequate spaces to make the tutorial look well structured.